+++ internal/features/students/events/read_model_event_handler.go
@@ -15,7 +15,7 @@ const StudentReadModelEventHandlerName = "student_read_model_event_handler"
 type StudentReadModelReader interface {
 	GetByID(ctx context.Context, studentID string) (*models.Student, error)
 	GetByUsername(ctx context.Context, username string) (*models.Student, error)
-	List(ctx context.Context) ([]models.Student, error)
+	List(ctx context.Context, opts ...ListOption) ([]models.Student, error)
 	ListByIEPServiceType(ctx context.Context, s string) ([]models.Student, error)
 }